# be/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Clinic Management API",
        "title": "Clinic Management",
        "contact": {},
        "version": "1.0"
    },
    "host": "clinicly.fly.dev",
    "basePath": "/api/v1",
    "paths": {
...
        "/medicine-report": {
            "get": {
                "description": "Get medicine report",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine report"
                ],
                "summary": "Get medicine report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Order by",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Order descending",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicine report response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineReportListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create medicine report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine report"
                ],
                "summary": "Create medicine report",
                "parameters": [
                    {
                        "description": "Medicine report data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicine report response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineReportResponse"
                        }
                    }
                }
            }
        },
        "/medicine-report/{id}": {
            "get": {
                "description": "Get medicine report by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine report"
                ],
                "summary": "Get medicine report by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine report id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicine report response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineReportResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update medicine report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine report"
                ],
                "summary": "Update medicine report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine report id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medicine report data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateMedicineReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicine report response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineReportResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete medicine report",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine report"
                ],
                "summary": "Delete medicine report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine report id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicine report response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineReportResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.MedicineReportListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MedicineReport"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controllers.MedicineReportRequest": {
            "type": "object",
            "required": [
                "medicine_id",
                "quantity"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "medicine_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "controllers.MedicineReportResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.MedicineReport"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "controllers.UpdateMedicineReportRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "medicine_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "models.Medicine": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
        "models.MedicineReport": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "medicine": {
                    "$ref": "#/definitions/models.Medicine"
                },
                "medicine_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
    }
}